Implemented Express Backend
Used a guide by Monique McIntyre on Medium to learn how to implement an Express
backend for a Vite app. Also, learned how to configure both Express and Vite to
implement a basic fetch command to ensure both server and client were working
and able to communicate with one another.